Created theme component
