+++ src/App.tsx
@@ -1,30 +1,13 @@
 import React from 'react';
-import '@fontsource/roboto'
-import { ThemeProvider, createMuiTheme } from '@material-ui/core/styles'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import useMediaQuery from '@material-ui/core/useMediaQuery'
-
-import Navbar from './components/Navbar/Navbar'
+import Theme from './components/Theme/Theme';
+import Navbar from './components/Navbar/Navbar';
 
 
 function App() {
-  const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
-
-  const theme = React.useMemo(
-    () =>
-      createMuiTheme({
-        palette: {
-          type: prefersDarkMode ? 'dark' : 'light',
-        },
-      }),
-    [prefersDarkMode],
-  )
-
   return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
+    <Theme>
       <Navbar title="Debt Calculator"></Navbar>
-    </ThemeProvider>
+    </Theme>
   );
 }
